requirements.txt
cryptography==44.0.2
pymysql==1.1.1
numpy==2.2.4
sqlalchemy==2.0.40
pandas==2.2.3
requests==2.32.3
textdistance==4.6.3
openpyxl==3.1.5
bleach==6.2.0
nltk==3.8.1
python-dotenv==1.1.0